log-alloc | Extract logical log serializer module
